+++ tokio-io/src/framed_write.rs
@@ -2,32 +2,11 @@ use std::io::{self, Read};
 use std::fmt;
 
 use {AsyncRead, AsyncWrite};
-use codec::Decoder;
+use codec::{Decoder, Encoder};
 use framed::Fuse;
 
 use futures::{Async, AsyncSink, Poll, Stream, Sink, StartSend};
 use bytes::BytesMut;
-
-/// Trait of helper objects to write out messages as bytes, for use with
-/// `FramedWrite`.
-pub trait Encoder {
-    /// The type of items consumed by the `Encoder`
-    type Item;
-
-    /// The type of encoding errors.
-    ///
-    /// `FramedWrite` requires `Encoder`s errors to implement `From<io::Error>`
-    /// in the interest letting it return `Error`s directly.
-    type Error: From<io::Error>;
-
-    /// Encodes a frame into the buffer provided.
-    ///
-    /// This method will encode `item` into the byte buffer provided by `dst`.
-    /// The `dst` provided is an internal buffer of the `Framed` instance and
-    /// will be written out when possible.
-    fn encode(&mut self, item: Self::Item, dst: &mut BytesMut)
-              -> Result<(), Self::Error>;
-}
 
 /// A `Sink` of frames encoded to an `AsyncWrite`.
 pub struct FramedWrite<T, E> {
